Rad insulation + Do after proc
Thanks a ton to Rohesie for helping me make this comply with TG standards.

+++ code/game/objects/structures/chess.dm
@@ -8,9 +8,11 @@
 	max_integrity = 100
 
 /obj/structure/chess/wrench_act(mob/user, obj/item/tool)
-	to_chat(user, "<span class='notice'>You take apart the chess piece.</span>")
-	var/obj/item/stack/sheet/metal/M = new (drop_location(), 2)
-	M.add_fingerprint(user)
+	to_chat(user, "<span class='notice'>You start to take apart the chess piece.</span>")
+	if(!do_after(user, 0.5 SECONDS, target = src))
+		return TRUE
+	var/obj/item/stack/sheet/metal/metal_sheets = new (drop_location(), 2)
+	metal_sheets.add_fingerprint(user)
 	tool.play_tool_sound(src)
 	qdel(src)
 	return TRUE
